beta 0.0.3

+++ include/msl/numerical.hpp
@@ -30,7 +30,7 @@ namespace msl
      * @param ApproximateValue The estimated or approximate value.
      * @return The absolute difference between TrueValue and ApproximateValue.
      */
-    constexpr float ATE(float TrueValue, float ApproximateValue) noexcept { return abs(TrueValue - ApproximateValue); }
+    constexpr double ATE(double TrueValue, double ApproximateValue) noexcept { return abs(TrueValue - ApproximateValue); }
 
     /**
      * @brief Calculates the Absolute Approximate Error (AAE) between two estimates.
@@ -42,7 +42,7 @@ namespace msl
      * @param OldEstimate The previous or old estimate value.
      * @return The absolute value of the difference between NewEstimate and OldEstimate.
      */
-    constexpr float AAE(float NewEstimate, float OldEstimate) noexcept { return abs(NewEstimate - OldEstimate); }
+    constexpr double AAE(double NewEstimate, double OldEstimate) noexcept { return abs(NewEstimate - OldEstimate); }
 
     /**
      * @brief Calculates the Relative True Error (RTE) between a true value and an approximate value.
@@ -52,9 +52,9 @@ namespace msl
      *
      * @param TrueValue The actual (true) value.
      * @param ApproximateValue The estimated or approximate value.
-     * @return The relative true error as a float.
+     * @return The relative true error as a double.
      */
-    constexpr float RTE(float TrueValue, float ApproximateValue) noexcept { return abs(ATE(TrueValue, ApproximateValue) / TrueValue); }
+    constexpr double RTE(double TrueValue, double ApproximateValue) noexcept { return abs(ATE(TrueValue, ApproximateValue) / TrueValue); }
 
     /**
      * @brief Calculates the Relative Approximate Error (RAE) between two estimates.
@@ -65,9 +65,9 @@ namespace msl
      *
      * @param NewEstimate The current (new) estimate value.
      * @param OldEstimate The previous (old) estimate value.
-     * @return The relative approximate error as a float.
+     * @return The relative approximate error as a double.
      */
-    constexpr float RAE(float NewEstimate, float OldEstimate) noexcept { return abs(AAE(NewEstimate, OldEstimate) / NewEstimate); }
+    constexpr double RAE(double NewEstimate, double OldEstimate) noexcept { return abs(AAE(NewEstimate, OldEstimate) / NewEstimate); }
 
     /**
      * @brief Calculates the approximate relative error threshold for a given number of significant digits.
@@ -76,37 +76,37 @@ namespace msl
      * the stopping criterion based on the number of significant digits.
      *
      * @param n Number of significant digits.
-     * @return The relative error threshold as a float.
+     * @return The relative error threshold as a double.
      */
-    constexpr float Es(int n) noexcept { return (0.5 * pow(10, 2 - n)); }
+    constexpr double Es(int n) noexcept { return (0.5 * pow(10, 2 - n)); }
 
     /**
-     * @brief Computes the square of a floating-point number.
+     * @brief Computes the square of a doubleing-point number.
      *
      * This function returns the result of raising the input value to the power of 2.
      *
-     * @param x The floating-point number to be squared.
+     * @param x The doubleing-point number to be squared.
      * @return The square of the input value.
      *
      * @note This function is marked as constexpr and noexcept.
      */
-    constexpr float square(float x) noexcept { return pow(x, 2); }
+    constexpr double square(double x) noexcept { return pow(x, 2); }
 
     /**
      * @brief Calculates the roots of a quadratic equation of the form ax^2 + bx + c = 0.
      *
      * This function uses the quadratic formula to find the real roots of the equation.
-     * The roots are returned as a std::vector<float> containing two elements.
+     * The roots are returned as a std::vector<double> containing two elements.
      * If the discriminant is negative, the results may be NaN due to sqrt of a negative number.
      *
      * @param a Coefficient of x^2.
      * @param b Coefficient of x.
      * @param c Constant term.
-     * @return std::vector<float> A vector containing the two roots.
+     * @return std::vector<double> A vector containing the two roots.
      */
-    std::vector<float> QuadEquRoots(float a, float b, float c) noexcept
+    std::vector<double> QuadEquRoots(double a, double b, double c) noexcept
     {
-        std::vector<float> roots;
+        std::vector<double> roots;
         roots.push_back((-b + sqrt(square(b) - (4 * a * c))) / (2 * a));
         roots.push_back((-b - sqrt(square(b) - (4 * a * c))) / (2 * a));
         return roots;
@@ -123,7 +123,7 @@ namespace msl
      * @param E The desired error tolerance.
      * @return The number of iterations required to achieve the specified precision.
      */
-    constexpr float n(float a, float b, float E) noexcept { return (log(b - a) - log(E)) / log(2); }
+    constexpr double n(double a, double b, double E) noexcept { return (log(b - a) - log(E)) / log(2); }
 
     /**
      * @brief Approximates the root of a continuous function using the Bisection method.
@@ -137,27 +137,27 @@ namespace msl
      * @param iterations   Maximum number of iterations to perform.
      * @param tolerance    Optional threshold to stop when the interval size or f(Xr) is below this value (default = 1e-5).
      * @param ShowIter     If true, prints each iteration step (default = false).
-     * @return float       The approximated root value, or NaN if the input is invalid or method fails.
+     * @return double       The approximated root value, or NaN if the input is invalid or method fails.
      */
-    float Bisection(std::function<float(float)> f, float Xl, float Xu, int iterations, float tolerance = 1e-5, bool ShowIter = false) noexcept
+    double Bisection(std::function<double(double)> f, double Xl, double Xu, int iterations, double tolerance = 1e-5, bool ShowIter = false) noexcept
     {
         if (Xl > Xu)
         {
             meta::lite::io::println_error("Error: Lower bound is greater than upper bound.");
             return NAN;
         }
-        float Fx_l = f(Xl);
-        float Fx_u = f(Xu);
+        double Fx_l = f(Xl);
+        double Fx_u = f(Xu);
         if (Fx_l * Fx_u > 0)
         {
             meta::lite::io::println_error("Error: No sign change in the interval. Bisection method may not converge.");
             return NAN;
         }
-        float Xr = Xl;
+        double Xr = Xl;
         for (int i = 0; i < iterations; ++i)
         {
             Xr = (Xl + Xu) / 2.0f;
-            float Fx_r = f(Xr);
+            double Fx_r = f(Xr);
             if (ShowIter)
             {
                 std::cout << "Iteration " << (i + 1)
@@ -194,15 +194,15 @@ namespace msl
      * @param iterations   The maximum number of iterations to perform.
      * @param tolerance    Optional stopping criterion; stops if |Xi+1 - Xi| < tolerance (default = 1e-5).
      * @param ShowIter     If true, prints each iteration step (default = false).
-     * @return float       The estimated root, or the last Xi if convergence fails.
+     * @return double       The estimated root, or the last Xi if convergence fails.
      */
-    float SecantMethod(std::function<float(float)> f, float Xi, float Xi_1, int iterations, float tolerance = 1e-5, bool ShowIter = false) noexcept
+    double SecantMethod(std::function<double(double)> f, double Xi, double Xi_1, int iterations, double tolerance = 1e-5, bool ShowIter = false) noexcept
     {
-        float Xi_new;
+        double Xi_new;
         for (int i = 0; i < iterations; ++i)
         {
-            float Fxi = f(Xi);
-            float Fxi_1 = f(Xi_1);
+            double Fxi = f(Xi);
+            double Fxi_1 = f(Xi_1);
             if (Fxi == Fxi_1)
             {
                 meta::lite::io::println_error("Division by zero in Secant Method. (f(Xi) == f(Xi-1))");
@@ -238,13 +238,13 @@ namespace msl
      * @param iteration Maximum number of iterations to perform.
      * @param ShowIter  Optional: show step-by-step details (default = false).
      * @param tolerance Optional: stop if |Xr - previous Xr| < tolerance (default = 1e-5).
-     * @return float    The approximated root value, or -1 if input is invalid.
+     * @return double    The approximated root value, or -1 if input is invalid.
      */
-    float False_Position(std::function<float(float)> f, float Xl, float Xu, int iteration, bool ShowIter = false, float tolerance = 1e-5) noexcept
+    double False_Position(std::function<double(double)> f, double Xl, double Xu, int iteration, bool ShowIter = false, double tolerance = 1e-5) noexcept
     {
-        float Xr = Xl; // initial guess (just for setup)
-        float Xr_old = Xr;
-        float Fx_l = f(Xl), Fx_u = f(Xu), Fx_r;
+        double Xr = Xl; // initial guess (just for setup)
+        double Xr_old = Xr;
+        double Fx_l = f(Xl), Fx_u = f(Xu), Fx_r;
         if (Xl > Xu)
         {
             meta::lite::io::println_error("Error: the lower value is greater than the upper value.");
@@ -310,15 +310,15 @@ namespace msl
      * @param iteration  Number of iterations to perform.
      * @param ShowIter   Optional: show steps during iterations (default = false).
      * @param tolerance  Optional: stop if |Xi+1 - Xi| < tolerance (default = 1e-5).
-     * @return float     The approximated root.
+     * @return double     The approximated root.
      */
-    float Newton_Raphson(std::function<float(float)> fx, std::function<float(float)> fd, float X_0, int iteration, bool ShowIter = false, float tolerance = 1e-5) noexcept
+    double Newton_Raphson(std::function<double(double)> fx, std::function<double(double)> fd, double X_0, int iteration, bool ShowIter = false, double tolerance = 1e-5) noexcept
     {
-        float Xi_old = X_0, Xi_new;
+        double Xi_old = X_0, Xi_new;
         for (int i = 0; i < iteration; i++)
         {
-            float Fi_old = fx(Xi_old);
-            float Fi_d = fd(Xi_old);
+            double Fi_old = fx(Xi_old);
+            double Fi_d = fd(Xi_old);
             if (Fi_d == 0)
             {
                 std::cerr << "Division by zero: f'(x) = 0 at iteration " << i + 1 << std::endl;
