+++ include/msl/statics.hpp
@@ -25,23 +25,23 @@ namespace msl
 
     const double PI(3.14159);
 
-    constexpr float square(float x) noexcept { return std::pow(x, 2); }
+    constexpr double square(double x) noexcept { return std::pow(x, 2); }
     
     // convert the degrees to radians
-    constexpr float toRadians(float degrees) noexcept { return degrees * PI / 180.0; }
+    constexpr double toRadians(double degrees) noexcept { return degrees * PI / 180.0; }
     
     // convert the randians to degrees
-    constexpr float todegrees(float radians) noexcept { return (radians * 180.0) / PI; }
+    constexpr double todegrees(double radians) noexcept { return (radians * 180.0) / PI; }
     
-    constexpr float CosineLaw(float A, float B, float angleDegrees) noexcept
+    constexpr double CosineLaw(double A, double B, double angleDegrees) noexcept
     {
-        float angleRad = toRadians(angleDegrees);
+        double angleRad = toRadians(angleDegrees);
         return sqrt(square(A) + square(B) - 2 * A * B * cos(angleRad));
     }
 
-    constexpr float SineLawFindSide(float B, float angleA, float angleB) noexcept { return (B * sin(toRadians(angleA))) / sin(toRadians(angleB)); }
+    constexpr double SineLawFindSide(double B, double angleA, double angleB) noexcept { return (B * sin(toRadians(angleA))) / sin(toRadians(angleB)); }
     
-    constexpr float thataSineLaw(float Fr, float F, float Thata) noexcept { return todegrees(asin((F * sin(toRadians(Thata))) / Fr)); }
+    constexpr double thataSineLaw(double Fr, double F, double Thata) noexcept { return todegrees(asin((F * sin(toRadians(Thata))) / Fr)); }
     
     /// @brief Class to handle Cartesian Vectors and Forces
     class CartesianVector
